backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
bcrypt==4.0.1
cryptography==43.0.0
psycopg2-binary==2.9.9
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
scikit-learn==0.24.2
joblib==1.0.1
requests==2.31.0